Backend/api/managers/requirements.txt
flask
flask_restplus
flask_cors
apis
Werkzeug
requests
neo4j-driver
blinker